requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# HTTP and File Handling
python-multipart==0.0.6
httpx==0.26.0
aiofiles==23.2.1

# AWS Integration
boto3==1.34.34
botocore==1.34.34
aiobotocore==2.11.2

# Configuration and Environment
python-dotenv==1.0.0
typing-extensions==4.9.0

# Data Validation and Processing
jsonschema==4.20.0
email-validator==2.1.0.post1

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Rate Limiting
slowapi==0.1.9

# Monitoring and Logging
prometheus-client==0.19.0
structlog==23.2.0
python-json-logger==2.0.7

# Database (if needed for future expansion)
sqlalchemy==2.0.25
alembic==1.13.1

# Additional Utilities
python-dateutil==2.8.2
pytz==2023.4
requests==2.31.0
